+++ out/src/StatusBarItem.js
@@ -16,10 +16,11 @@ class StatusBarItem {
             this.statusBarItem.hide();
             return;
         }
-        // Only update status if an HTML file
+        // Update status if an HTML file:
         if (this.utilities.checkDocumentIsHTML(false)) {
             this.statusBarItem.text = Constants.ExtensionConstants.STATUS_BAR_HTML_TEXT;
             this.statusBarItem.show();
+            // Update status if a JS file:
         }
         else if (this.utilities.getDocumentType() === "javascript") {
             this.statusBarItem.text = Constants.ExtensionConstants.STATUS_BAR_JS_TEXT;