+++ out/src/extension.js
@@ -13,6 +13,9 @@ function activate(context) {
     let disposableStatusBar = vscode.window.onDidChangeActiveTextEditor(statusBarItem.updateStatusbar, statusBarItem, context.subscriptions);
     let previewUri = vscode.Uri.parse(Constants.ExtensionConstants.PREVIEW_URI);
     // Register the commands that are provided to the user
+    let disposableSideJSFlowChart = vscode.commands.registerCommand('extension.sideJSFlowChart', () => {
+        utilities.initJSFlowChart(vscode.ViewColumn.Two);
+    });
     let disposableSidePreview = vscode.commands.registerCommand('extension.sidePreview', () => {
         utilities.initHTMLPreview(vscode.ViewColumn.Two);
     });
@@ -25,9 +28,12 @@ function activate(context) {
         }
     });
     // push to subscriptions list so that they are disposed automatically
+    // HTML Preview:
     context.subscriptions.push(disposableSidePreview);
     context.subscriptions.push(disposableStandalonePreview);
     context.subscriptions.push(disposableInBrowser);
+    // JS FlowChart:
+    context.subscriptions.push(disposableSideJSFlowChart);
 }
 exports.activate = activate;
 // This method is called when extension is deactivated