+++ src/StatusBarItem.ts
@@ -9,10 +9,12 @@ export default class StatusBarItem {
     utilities: Utilities;
 
     constructor(utilities?: Utilities) {
+
         this.statusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left);
         this.statusBarItem.command = "extension.sidePreview";
         this.statusBarItem.tooltip = Constants.ExtensionConstants.STATUS_BAR_HTML_TOOLTIP;
         this.utilities = utilities && utilities || new Utilities();
+
     }
 
     updateStatusbar() {
@@ -22,12 +24,13 @@ export default class StatusBarItem {
             this.statusBarItem.hide();
             return;
         }
-        // Only update status if an HTML file
+        // Update status if an HTML file:
         if ( this.utilities.checkDocumentIsHTML(false) ) {
 
             this.statusBarItem.text = Constants.ExtensionConstants.STATUS_BAR_HTML_TEXT;
             this.statusBarItem.show();
 
+        // Update status if a JS file:
         } else if ( this.utilities.getDocumentType() === "javascript" ){
 
             this.statusBarItem.text = Constants.ExtensionConstants.STATUS_BAR_JS_TEXT;